Downloads/CS-Girls-Fall-Hack-ilyoumi/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
python-dotenv
authlib
httpx
itsdangerous
starlette
transformers 
sentence-transformers 
pdfplumber 
docx2txt 
google-generativeai